==> notebooks/00_exploration.ipynb <==
# <center>  **Data Exploration**

------

# Make Constant Variables For Data Directory to access Data Faster in Notebook
DATA_PATH_CHILD  = '../data/raw/G2000_data.csv'
DATA_PATH_PARENT = '../data/raw/EthA_data.csv'

##  Import libraries

import pandas as pd 
# import numpy as np 
pd.options.display.max_rows = 1000

## Child Data

# Read data of child and explore 
raw_df_child = pd.read_csv(DATA_PATH_CHILD)

%store raw_df_child

raw_df_child.sample(10)

- `Allele : is a variant of the sequence of nucleotides at a particular location, or locus, on a DNA molecule`
- visualize whats going on and what alleles are ?
- https://learn.genetics.utah.edu/content/genetics/whatallele

raw_df_child.info()

- `Name : is the generation of the individual with an Unique identifier `

- `Gender : is the sex of the individual (M or F)`

- `ParentM : is the mother of the individual`

- `ParentF : is the father of the individual`

- `EthA : is the ethnicity of the individual of Africa American`

- `EthE : is the ethnicity of the individual of Estonian`

- `EthK : is the ethnicity of the individual of Korean`

- `EthP : is the ethnicity of the individual of the Palestinian`

- `Allele1 : is a variant of the sequence of nucleotides at a particular location, or locus, on a DNA molecule`

- `Allele2 : is a variant of the sequence of nucleotides at a particular location, or locus, on a DNA molecule`

raw_df_child.isnull().sum()

allele1_information_child = raw_df_child['Allele1'][0]
len(allele1_information_child)

#allele1_information_child.to_csv('allele1_information_child.csv')

## **Explore more about child data**

### Test-a

# raw_df_child['Allele1'] = raw_df_child['Allele1'].apply(lambda x : ''.join(x[:1000]))

# raw_df_child['Allele2'] = raw_df_child['Allele2'].apply(lambda x : ''.join(x[:1000]))

# raw_df_child['Full_seq_dna']= raw_df_child['Allele1'] + raw_df_child['Allele2']

# raw_df_child

len(raw_df_child['Allele1'][16927])

test_parent_1 = raw_df_child [raw_df_child['ParentM'] == 'A8184']
test_parent_1

### Test-b

# # Display information of child of one Parent  (mother)
# test_parent_2 = raw_df_child [raw_df_child['ParentM'] == 'A4285']
# test_parent_2

# first_test_child_2 = test_parent_2['Full_seq_dna'][5883]
# # same first_test_child dna data as csv 
# len(first_test_child_2)

# # 1 - 5880
# # 2 - 5881
# # 3 - 5882
# # 4 - 5883

# second_test_child_2 = test_parent_2['Full_seq_dna'][8291]
# # same first_test_child dna data as csv 
# len(second_test_child_2)

# # 1 - 8288
# # 2 - 8289
# # 3 - 8290
# # 4 - 8291

# with open('childG20008291_Same_parentM.txt', 'w') as file:
#     file.write(second_test_child_2)

- ` here we can see that data for parent has  A8184 has 8 records`

------

## Parent Data

raw_df_parent = pd.read_csv(DATA_PATH_PARENT)

%store raw_df_parent

raw_df_parent.sample(10)

----

- `ParentM and ParentF is Na for 1st generation individuals `

----

raw_df_parent.info()

raw_df_child.isnull().sum()

allele1_information_parent = raw_df_parent['Allele1'][0]

allele1_information_parent

len(allele1_information_parent)

allele2_information_parent = raw_df_parent['Allele2'][0]

len(allele2_information_parent)

raw_df_parent [raw_df_parent['Name'] == 'A13904']
