==> notebooks/01_preparation.ipynb <==
# <center> Data Preparation </center>
## <center> Step 1 <center>

------

# Constants
DATA_PATH_CHILD  = '../data/raw/G2000_data.csv'
DATA_PATH_PARENT = '../data/raw/EthA_data.csv'
EXPORT_PATH = '../data/processed/1_first_processed_merged_df.pkl'

import pandas as pd 

raw_df_child = pd.read_csv(DATA_PATH_CHILD)

df_child = raw_df_child.copy()

raw_df_parent = pd.read_csv(DATA_PATH_PARENT)

df_parent = raw_df_parent.copy()

-----

<center> - Creating a new column in the DataFrame and assigning values based on concatenating the values from two existing columns <center>

<center>** full sequence of alleles **</center>

-----

## For child

df_child.sample(3)

- length of the sequence before applying the function to take the first 1000 bases of the sequence from allele 1

# Check the length of the sequence before truncating to 1000 bases
len(df_child['Allele1'].iloc[0])

# Truncate Allele sequences to the first 1000 bases and concatenate
df_child['Allele1'] = df_child['Allele1'].apply(lambda x : ''.join(x[:1000]))
df_child['Allele2'] = df_child['Allele2'].apply(lambda x : ''.join(x[:1000]))
df_child['Full_seq_dna_child']= df_child['Allele1'] + df_child['Allele2']

- length of the sequence after applying the function to take the first 1000 bases of the sequence

len(df_child['Allele1'].iloc[0])

df_child

Full sequence of dna after applying the function to take the first 1000 bases of the sequence in allel 1 and 2 respectively
- Length of Full sequence of DNA of child after concatenating allel 1 and 2 with 1000 bases from each one 

# Check the length of the full concatenated sequence for child
len(df_child['Full_seq_dna_child'].iloc[0])

## For parent 

- same process we made for child 

# Display a sample of the parent DataFrame
df_parent.sample(3)

- length of the sequence before applying the function to take the first 1000 bases of the sequence from allel 1

# Check the length of the sequence before truncating to 1000 bases
len(df_parent['Allele1'].iloc[0])

# Truncate Allele sequences for parent to the first 1000 bases and concatenate
df_parent['Allele1'] = df_parent['Allele1'].apply(lambda x : ''.join(x[:1000]))
df_parent['Allele2'] = df_parent['Allele2'].apply(lambda x : ''.join(x[:1000]))
df_parent['Full_seq_dna_parent']= df_parent['Allele1'] + df_parent['Allele2']

- length of the sequence of parent after applying the function to take the first 1000 bases of the sequence

# Check the length of the sequence after truncating to 1000 bases
len(df_parent['Allele1'].iloc[0])

Full sequence :
- Length of Full sequence of DNA of parent after concatenating allel 1 and 2 with 1000 bases from each one 

# Check the length of the full concatenated sequence for parent
len(df_parent['Full_seq_dna_parent'].iloc[0])

## Preparing data for tests

# Prepare data for tests
# make copy of G2000 ['ParentM','ParentF','Child_full_DNA_Seq']
df_child_data_copy = df_child[['ParentM','ParentF','Full_seq_dna_child']]

# Display information of (df_child_data_copy) G2000 ['ParentM','ParentF','Child_full_DNA_Seq']
df_child_data_copy.sample(10)

# make copy of EthA ['Name','Parent_full_DNA_Seq']
df_parent_data_copy = df_parent[['Name','Full_seq_dna_parent']]

df_parent_data_copy.sample(10)

# import re 
# df_test = pd.read_csv('test_child.csv')
# def remove_numbers(text):
#     return re.sub(r'\d+', '', text)
# # Apply the function to the 'Parent_full_DNA_Seq' column
# df_test['Parent_full_DNA_Seq'] = df_test['Parent_full_DNA_Seq'].apply(remove_numbers)
# # Print the modified DataFrame
# print(df_test)

## Extracting DNA sequences of parents for tests

### Test-a


# Extract DNA sequences for specific parents for tests
df_parent_data_copy[df_parent_data_copy['Name'] == 'A8184']

first_parent = df_parent_data_copy[df_parent_data_copy['Name'] == 'A8184']['Full_seq_dna_parent'].iloc[0]

len(first_parent)

# with open('ParentM_A8184.txt', 'w') as file:
#     file.write(first_parent)

df_parent_data_copy[df_parent_data_copy['Name'] == 'A13904']

second_parent = df_parent_data_copy[df_parent_data_copy['Name'] == 'A13904']['Full_seq_dna_parent'].iloc[0]

len(second_parent)

# with open('ParentM_A13904.txt', 'w') as file:
#     file.write(second_parent)

df_parent_data_copy[df_parent_data_copy['Name'] == 'A4285']

Third_parent_b = df_parent_data_copy[df_parent_data_copy['Name'] == 'A14108']['Full_seq_dna_parent'].iloc[0]

len(Third_parent_b)

# with open('ParentF_A14108.txt', 'w') as file:
#     file.write(Third_parent_b)

### Test-b

# Third_parent = df_parent_data_copy[df_parent_data_copy['Name'] == 'A4285']['Full_seq_dna'].iloc[0]

# with open('ParentM_A4285.txt', 'w') as file:
#     file.write(Third_parent)

# fourth_parent = df_parent_data_copy[df_parent_data_copy['Name'] == 'A4427']['Full_seq_dna'].iloc[0]

# with open('ParentF_A4427.txt', 'w') as file:
#     file.write(Third_parent)

# fifth_parent = df_parent_data_copy[df_parent_data_copy['Name'] == 'A14919']['Full_seq_dna'].iloc[0]

# with open('ParentF_A14919.txt', 'w') as file:
#     file.write(fifth_parent)

-----

## **Merging Operation**

-----

**Merging :**

Merge the dataframes using the `Name` column as the key. The key column for merging is located in the `df_parent` dataframe.



**Keys in `df_child` DataFrame:**

- Left DataFrame Key (`left_on`): `Name`
- Right DataFrame Key (`right_on`): `Father or Mother`



------

# Merging Once with Father and Once with Mother
parent_child_merge_father = pd.merge(df_parent_data_copy , df_child_data_copy , left_on = 'Name' , right_on= 'ParentF')
parent_child_merge_mother = pd.merge(df_parent_data_copy , df_child_data_copy , left_on = 'Name' , right_on= 'ParentM')

# check each dataframe
parent_child_merge_father.sample(10 , random_state= 42)

-----

- ` here when we check example row 1 we see that Name = 'A11756' and ParentF = 'A11756' and ParentM = 'A9729'`

- ` From this we can say that they are the same person`

- ` From data source they say that Name is unique identifier for a person`

- ` Finally, we conclude that Name is Father of the child ` 

------

parent_child_merge_mother

- For Father:

# Clean up and combine the merged dataframes
cleaned_data_first_part = parent_child_merge_father.drop(columns=['ParentM', 'ParentF'])
cleaned_data_first_part

first_merged_df= pd.concat([cleaned_data_first_part, parent_child_merge_mother])

-----

## Verify results 

-----

# Display a sample of the final merged DataFrame
first_merged_df.sample(10)

-----

# Export Data 

-----

first_merged_df.to_pickle(EXPORT_PATH)
