==> notebooks/02_feature_engineering.ipynb <==
# <center> Feature Engineering </center>

## Problem Statement  
 - The objective of this notebook is to perform feature extraction from the given dataset. Specifically, we aim to create a  Global alignment attempts to align two sequences from end to end and generate a Target Label Feature based on the Pairwise Alignment Score. 

## Feature Creation Steps

### 1. Creating a Global Weighted Alignment Sequence Feature

- By implementing the **needleman_wunsch_similarity** function to compare DNA sequences and calculate the score of paternity based on matching alleles using the Needleman-Wunsch algorithm. 
 
### 2. Generating a Target Label Feature
- By utilizing the Alignment sequence feature to generate a binary target label (0, 1). The label is determined by a predefined threshold based on medical research. 

------

# Constants
DATA_PATH = '../data/processed/1_first_processed_merged_df.pkl'
EXPORT_PATH = '../data/processed/2_second_processed_merged_df.pkl'

import pandas as pd 

## Read data

first_processed_merged_df = pd.read_pickle(DATA_PATH)

first_processed_merged_df_copy = first_processed_merged_df.copy()

first_processed_merged_df_copy.sample(10)

first_processed_merged_df.info()

------

## **Global Weighted Alignment Sequence Ratio**

### Compares two sequences (alleles) for a paternity test based on the Pairwise Alignment Score

- Global alignment aims to align the entire length of two sequences. It is suitable when the sequences are of similar length and are expected to be homologous throughout.
 

------

from Bio import pairwise2

def needleman_wunsch_similarity(seq1, seq2, match_score=3, mismatch_penalty=-1, gap_penalty=-2):
    alignments = pairwise2.align.globalms(seq1, seq2, match_score, mismatch_penalty, gap_penalty, gap_penalty)
    """
    Calculates the similarity score between two sequences using the Needleman-Wunsch algorithm.

    Parameters:
        seq1 (str): First DNA sequence.
        seq2 (str): Second DNA sequence.
        match_score (int, optional): Score for matches. Defaults to 3.
        mismatch_penalty (int, optional): Penalty for mismatches. Defaults to -1.
        gap_penalty (int, optional): Penalty for gaps. Defaults to -2.

    Returns:
        float: Similarity score between the two sequences.
    """
    if not alignments:
        return 0.0

    best_alignment = alignments[0]  # Highest score alignment 
    alignment_score = best_alignment[2]  # Score of the best alignment
    max_score = max(len(seq1), len(seq2)) * match_score
    similarity_score = alignment_score / max_score
    return similarity_score

# Example usage:
father_sequence = first_processed_merged_df_copy.iloc[20, 1]
child_sequence = first_processed_merged_df_copy.iloc[20, 2]
similarity_score = needleman_wunsch_similarity(father_sequence, child_sequence)
print("Similarity Score: {:.2%}".format(similarity_score))

How similar is the DNA of siblings?
- The concept of ‘shared’ DNA differs depending on how much DNA is being compared. When comparing our whole genome, all 3 billion letters, we share about 99.9 percent with every other human. However, we can also analyse shared DNA by narrowing our scope and just comparing the 3 million genetic differences we know exist between individuals. Of these 3 million differences, on average we share about 50 percent of those with our full siblings.

- Children inherit half of their DNA from their mother and half from their father. However, unless they are identical twins, siblings won’t inherit exactly the same DNA.
Depending on their biological sex, a parent will produce either a sperm or egg cell by meiosis. Meiosis is a process where a single cell divides twice to produce four cells containing half the original amount of genetic information. At fertilisation, two of these cells (one egg cell and one sperm cell) will unite to form a single cell. This cell will contain a full set of 46 chromosomes, which make up the child’s DNA.


------

# Constants for needleman_wunsch_similarity
# Still point of researching 
# need to be discussed
NEEDLEMAN_SIMILARITY = 0.95

## **Target Label**

def get_target(df):
    df['similarity_score'] = df.apply(lambda row: needleman_wunsch_similarity(row['Full_seq_dna_parent'], row['Full_seq_dna_child']), axis=1)
    df['target'] = (df['similarity_score'] >= NEEDLEMAN_SIMILARITY).astype(int)  # Set target to 1 or 0
    return df

second_processed_merged_df = get_target(first_processed_merged_df_copy)

second_processed_merged_df

second_processed_merged_df['target'].value_counts()

second_processed_merged_df.sample(100)

second_processed_merged_df.to_pickle(EXPORT_PATH)
