==> notebooks/03_augmentation.ipynb <==
# <center> Data Augmentation </center>
-----
## <center> Step 2 <center>

-------

- we need to load more data for parents that are not in our dataset to overcome the limitation of our dataset also underfitting the model

# Constants
DATA_PATH_RAW = '../data/raw/EthE_data.csv'
DATA_PATH_PROCESSED = '../data/processed/2_second_processed_merged_df.pkl'
EXPORT_PATH = '../data/processed/final_processed_data.pkl'

import pandas as pd 

e_raw_df = pd.read_csv(DATA_PATH_RAW)
e_raw_df.sample(4)

e_raw_df.info()

# first for child
e_raw_df['Allele1'] = e_raw_df['Allele1'].apply(lambda x : ''.join(x[:1000]))
e_raw_df['Allele2'] = e_raw_df['Allele2'].apply(lambda x : ''.join(x[:1000]))
e_raw_df['Full_seq_dna_parent']= e_raw_df['Allele1'] + e_raw_df['Allele2']

e_raw_df.sample(2)

e_raw_df_copy = e_raw_df[['Name', 'Full_seq_dna_parent']]

e_raw_df_copy['Full_seq_dna_parent'][20]

e_raw_df_copy['target'] = 0 

e_raw_df_copy.sample(4)

second_processed_merged_df = pd.read_pickle(DATA_PATH_PROCESSED)

second_processed_merged_df

second_processed_merged_df.tail(5)

second_processed_merged_df.drop(second_processed_merged_df[second_processed_merged_df['target'] == 1].sample(15500).index, inplace=True)

# second_processed_merged_df = second_processed_merged_df.drop(second_processed_merged_df.index[range(5)])

# second_processed_merged_df.drop(second_processed_merged_df[second_processed_merged_df['target'] == 1].sample(2000).index, inplace=True)

second_processed_merged_df.drop(second_processed_merged_df[second_processed_merged_df['target'] == 1].sample(184).index, inplace=True)

second_processed_merged_df

final_processed_data = pd.concat([second_processed_merged_df , e_raw_df_copy])

final_processed_data

df_pre = final_processed_data[['Full_seq_dna_parent', 'Full_seq_dna_child']]
df_pre

final_processed_data['target'].value_counts()

final_processed_data['target'].value_counts()

# read first Parent_full_DNA_Seq column
#final_processed_data[['Parent_full_DNA_Seq', 'Child_full_DNA_Seq']].sample(4)

#final_processed_data.iloc[0]  # first row
# final_processed_data.iloc[1] # second row
# final_processed_data.iloc[2][2] # third row
# final_processed_data.iloc[3][2] # fourth row    
# final_processed_data.iloc[4][2] # fifth row
# final_processed_data.iloc[5][2] # sixth row

# final_processed_data.iloc[6][2]

# final_processed_data.iloc[7][2]
# final_processed_data.iloc[8][2]
final_processed_data.iloc[11][2]

final_processed_data.to_pickle(EXPORT_PATH)
